pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metricflow-server"
version = "0.1.0"
description = "Self-hosted REST API for dbt's MetricFlow semantic layer"
license = { file = "LICENSE" }
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "pydantic-settings>=2.0",
    "metricflow>=0.207.3,<0.209.0",
    "dbt-metricflow>=0.8.2,<0.9.0",
]

[project.optional-dependencies]
bigquery = ["dbt-bigquery>=1.7"]
redshift = ["dbt-redshift>=1.7"]

[project.scripts]
metricflow-server = "metricflow_server.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/metricflow_server"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "pytest-mock>=3.14",
]

